Fixing the packaging
