+++ crypt4gh/keys/__main__.py
@@ -9,7 +9,6 @@ assert sys.version_info >= (3, 6), "This tool requires python version 3.6 or hig
 import os
 import io
 import logging
-from base64 import b64decode, b64encode
 from functools import partial
 from getpass import getpass
 from pathlib import Path
@@ -17,7 +16,7 @@ from pathlib import Path
 from docopt import docopt
 
 from .. import __title__, __version__, PROG
-from . import ssh, c4gh
+from . import c4gh, get_public_key, get_private_key
 
 LOG = logging.getLogger(__name__)
 
@@ -58,77 +57,6 @@ Environment variables:
  
 '''
 
-#######################################################################
-## Loading
-#######################################################################
-
-
-def load_from_pem(filepath):
-    with open(filepath, 'rb') as f:
-        lines = []
-
-        # Strip empty lines and newline characters
-        for l in f.readlines():
-            l = l.strip()
-            if l:
-                lines.append(l)
-
-        if (not lines
-            or not lines[0].startswith(b'-----BEGIN ')
-            or not lines[-1].startswith(b'-----END ')
-            ):
-            raise ValueError('Not a PEM format')
-
-        return b64decode(b''.join(lines[1:-1]))
-
-def get_public_key(filepath):
-    '''Read the public key from keyfile location.'''
-
-    with open(filepath, 'rb') as f:
-        lines = []
-
-        # Strip empty lines and newline characters
-        for l in f.readlines():
-            l = l.strip()
-            if l:
-                lines.append(l)
-
-        if not lines:
-            raise ValueError('Empty key')
-
-        line = lines[0]
-
-        if b'CRYPT4GH' in line: # it's Crypt4GH key
-            LOG.info('Loading a Crypt4GH public key')
-            return b64decode(b''.join(lines[1:-1]))
-
-        if line[:4] == b'ssh-': # It's an SSH key
-            LOG.info('Loading an OpenSSH public key')
-            return ssh.get_public_key(line)
-
-    raise NotImplementedError('Unsupported key format')
-
-
-def get_private_key(filepath, callback):
-    '''Read the private key from keyfile location.
-
-    If the private key is encrypted, the user will be prompted for the passphrase.
-    '''
-    data = load_from_pem(filepath)
-    stream = io.BytesIO(data)
-    magic_word = stream.read(len(c4gh.MAGIC_WORD)) # start with C4GH, it's smaller
-
-    if magic_word == c4gh.MAGIC_WORD: # It's a Crypt4GH key
-        LOG.info('Loading a Crypt4GH private key')
-        return c4gh.parse_private_key(stream, callback)
-
-    magic_word += stream.read(len(ssh.MAGIC_WORD)-len(c4gh.MAGIC_WORD))
-    if magic_word == ssh.MAGIC_WORD: # It's an SSH key
-        LOG.info('Loading an OpenSSH private key')
-        return ssh.parse_private_key(stream, callback)[0] # we also return the pubkey
-    
-    raise ValueError('Invalid key format')
-
 
 ###################
 ### CLI
@@ -159,32 +87,32 @@ def _main():
     seckey = Path(args['--sk']).expanduser()
     comment = args['-C'].encode() if args['-C'] else None
 
-    comment_message = f" (for {args['-C']})" if comment else ""
+    comment_message = f" (for {args['-C']})" if comment else ''
 
     if args['--relock']:
 
-        print(f"Relocking secret key pair{comment_message}.")
+        print(f'Relocking secret key {args["--sk"]}{comment_message}.')
 
         if not seckey.exists():
             print(f'Secret key {seckey} not found', file=sys.stderr)
-            sys.exit(1)
-
-        if args['--nocrypt']:
-            print("Can't both --relock and --nocrypt", file=sys.stderr)
             sys.exit(1)
 
         old_passphrase = os.getenv('C4GH_PASSPHRASE')
         cb = partial(getpass, prompt=f'Passphrase for {args["--sk"]}: ') if old_passphrase is None else lambda: old_passphrase
         seckey_decrypted = get_private_key(str(seckey), cb)
 
-        passphrase1 = getpass(prompt=f'Enter new passphrase for {args["--sk"]} (empty for no passphrase): ').encode()
-        passphrase2 = getpass(prompt=f'Enter new passphrase for {args["--sk"]} (again): ').encode()
-        if passphrase1 != passphrase2:
+        passphrase1 = passphrase2 = None
+        if not args['--nocrypt']:
+            passphrase1 = getpass(prompt=f'Enter new passphrase for {args["--sk"]} (empty for no passphrase): ').encode()
+            passphrase2 = getpass(prompt=f'Enter new passphrase for {args["--sk"]} (again): ').encode()
+
+        if passphrase1 != passphrase2: # including None=None
             print('Passphrases do not match', file=sys.stderr)
             sys.exit(1)
-        # We don't check if old_passphrase is the same as passphrase1/2        
+            # We don't check if old_passphrase is the same as passphrase1/2
+
         c4gh.relock(seckey, seckey_decrypted, passphrase1, comment)
-        print("Your private key has been relocked in", seckey)
+        print('Your private key has been relocked in', seckey)
         return
 
 
@@ -217,15 +145,15 @@ def _main():
             k.parent.mkdir(mode=0o700, parents=True, exist_ok=True)
 
     # ... and cue music
-    c4gh.generate(seckey, pubkey, passphrase=passphrase1, comment=comment)
-    print("Your private key has been saved in", seckey)
-    print("Your public key has been saved in", pubkey)
+    c4gh.generate(seckey, pubkey, passphrase1, comment)
+    print('Your private key has been saved in', seckey)
+    print('Your public key has been saved in', pubkey)
 
 def main():
     try:
         _main()
     except KeyboardInterrupt:
-        print("... Interrupted", file=sys.stderr)
+        print('... Interrupted', file=sys.stderr)
         sys.exit(1)
